util: move get_run_directory to new file
